+++ nengo_edge/tests/test_tflite_runner.py
@@ -8,7 +8,7 @@ import tensorflow as tf
 from nengo_edge_hw import coral
 from nengo_edge_models.kws.tests.test_models import _test_lmu
 
-from nengo_edge.tflite_runner import Runner
+from nengo_edge.tflite_runner import TFLiteRunner
 
 
 @pytest.mark.parametrize("return_sequences", (True, False))
@@ -36,8 +36,8 @@ def test_runner_streaming(
 
     inputs = rng.uniform(-0.5, 0.5, size=(batch_size, 16000))
 
-    runner = Runner(tmp_path / "exported")
-    runner_batch = Runner(tmp_path / "exported-batch")
+    runner = TFLiteRunner(tmp_path / "exported")
+    runner_batch = TFLiteRunner(tmp_path / "exported-batch")
 
     # check that batched and non-batched models produce the same output (state is being
     # properly reset between batch items)
@@ -82,10 +82,12 @@ def test_runner_errors(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
     model_desc = _test_lmu()
     model_desc.n_unroll = 2
     monkeypatch.setattr(coral.host.Interface, "io_dtype", None)
-    host = coral.host.Interface(model_desc, build_dir=tmp_path, use_device=False)
-    host.export_model(tmp_path, include_feature_extractor=False)
+    host = coral.host.Interface(
+        model_desc, build_dir=tmp_path, use_device=False, mode="model-only"
+    )
+    host.export_model(tmp_path)
 
-    runner = Runner(tmp_path)
+    runner = TFLiteRunner(tmp_path)
 
     with pytest.raises(ValueError, match="evenly divided by unroll"):
         runner.run(np.zeros((1, 3, 10)))
@@ -106,7 +108,7 @@ def test_runner_quantized(tmp_path: Path, rng: np.random.RandomState) -> None:
     )
     host.export_model(tmp_path)
 
-    runner = Runner(tmp_path)
+    runner = TFLiteRunner(tmp_path)
 
     x = rng.uniform(-1, 1, size=(1, 16000))
     y0 = runner.run(x)